+++ tests/integration/integrations/test_oauth.py
@@ -9,6 +9,7 @@ import re
 
 import jubilant
 import pytest
+import requests
 from playwright.sync_api import expect, sync_playwright
 
 from tests.integration.types import App
@@ -19,7 +20,9 @@ logger = logging.getLogger(__name__)
 @pytest.mark.parametrize(
     "app_fixture, endpoint",
     [
+        ("spring_boot_app", "oauth2/authorization/oidc"),
         ("flask_app", "login"),
+        ("django_app", "auth_login"),
     ],
 )
 def test_oauth_integrations(
@@ -28,6 +31,8 @@ def test_oauth_integrations(
     endpoint,
     request: pytest.FixtureRequest,
     identity_bundle,
+    browser_context_manager,
+    http: requests.Session,
 ):
     """
     arrange: set up the test Juju model and deploy the workload charm.
@@ -42,15 +47,22 @@ def test_oauth_integrations(
     app = request.getfixturevalue(app_fixture)
     status = juju.status()
 
-    if not status.apps.get(app.name).relations.get("oidc"):
-        juju.integrate(f"{app.name}", "hydra")
-
     if not status.apps.get(app.name).relations.get("ingress"):
         juju.integrate(f"{app.name}", "traefik-public")
 
     juju.wait(
         jubilant.all_active,
-        timeout=30 * 60,
+        timeout=10 * 60,
+        delay=5,
+    )
+
+    if not status.apps.get(app.name).relations.get("oidc"):
+        juju.integrate(f"{app.name}", "hydra")
+
+    juju.wait(
+        jubilant.all_active,
+        timeout=10 * 60,
+        delay=5,
     )
 
     if not _admin_identity_exists(juju, test_email):
@@ -69,15 +81,22 @@ def test_oauth_integrations(
         secret_id = [secret for secret in secrets if secrets[secret].get("name") == test_secret][0]
 
     juju.cli("grant-secret", secret_id, "kratos")
-    juju.run(
+    result = juju.run(
         "kratos/0",
         "reset-password",
         {"email": test_email, "password-secret-id": secret_id.split(":")[-1]},
     )
+    logger.info("results reset-password %s", result.results)
 
     res = json.loads(
         juju.run("traefik-public/0", "show-proxied-endpoints").results["proxied-endpoints"]
     )
+    logger.info("result show-proxied %s", res)
+
+    # make sure the app is alive
+    response = http.get(res[app.name]["url"], timeout=5, verify=False)
+    assert response.status_code == 200
+
     _assert_idp_login_success(res[app.name]["url"], endpoint, test_email, test_password)
 
 
@@ -91,15 +110,15 @@ def _admin_identity_exists(juju, test_email):
 
 
 def _assert_idp_login_success(app_url: str, endpoint: str, test_email: str, test_password: str):
-
     with sync_playwright() as p:
         browser = p.chromium.launch(headless=True)
         context = browser.new_context(ignore_https_errors=True)
         page = context.new_page()
         page.goto(f"{app_url}/{endpoint}")
+        logger.info("Page content: %s", page.content())
         expect(page).not_to_have_title(re.compile("Sign in failed"))
         page.get_by_label("Email").fill(test_email)
         page.get_by_label("Password").fill(test_password)
         page.get_by_role("button", name="Sign in").click()
         expect(page).to_have_url(re.compile(f"^{app_url}/profile.*"))
-        expect(page.get_by_role("heading", name=f"Welcome, {test_email}!")).to_be_visible()
+        assert f"Welcome, {test_email}!" in page.content()