+++ tests/unit/general/test_oauth.py
@@ -14,7 +14,9 @@ import pytest
 from conftest import OAUTH_RELATION_DATA_EXAMPLE
 from ops import testing
 
+from examples.django.charm.src.charm import DjangoCharm
 from examples.flask.charm.src.charm import FlaskCharm
+from examples.springboot.charm.src.charm import SpringBootCharm
 from paas_charm.utils import config_metadata
 
 
@@ -26,8 +28,8 @@ from paas_charm.utils import config_metadata
             FlaskCharm,
             "flask",
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "openid profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
             },
             "/bin/python3 -m gunicorn -c /flask/gunicorn.conf.py app:app -k [ sync ]",
             {
@@ -47,6 +49,104 @@ from paas_charm.utils import config_metadata
                 "FLASK_OIDC_JWKS_URL": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
             },
             id="flask-oidc",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            "django",
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            "/bin/python3 -m gunicorn -c /django/gunicorn.conf.py django_app.wsgi:application -k [ sync ]",
+            {
+                "DJANGO_OIDC_REDIRECT_PATH": "/oauth/callback",
+                "DJANGO_OIDC_SCOPES": "openid profile email phone",
+                "DJANGO_BASE_URL": "http://juju.test/",
+                "DJANGO_SECRET_KEY": "test",
+                "DJANGO_ALLOWED_HOSTS": '["juju.test"]',
+                "DJANGO_PEER_FQDNS": "django-k8s-0.django-k8s-endpoints.test-model.svc.cluster.local",
+                "DJANGO_OIDC_CLIENT_ID": "test-client-id",
+                "DJANGO_OIDC_CLIENT_SECRET": "abc",
+                "DJANGO_OIDC_API_BASE_URL": "https://traefik_ip/model_name-hydra",
+                "DJANGO_OIDC_AUTHORIZE_URL": "https://traefik_ip/model_name-hydra/oauth2/auth",
+                "DJANGO_OIDC_ACCESS_TOKEN_URL": "https://traefik_ip/model_name-hydra/oauth2/token",
+                "DJANGO_OIDC_USER_URL": "https://traefik_ip/model_name-hydra/userinfo",
+                "DJANGO_OIDC_CLIENT_KWARGS": '{"scope": "openid profile email phone"}',
+                "DJANGO_OIDC_JWKS_URL": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
+                "POSTGRESQL_DB_CONNECT_STRING": "postgresql://test-username:test-password@test-postgresql:5432/django-k8s",
+                "POSTGRESQL_DB_FRAGMENT": "",
+                "POSTGRESQL_DB_HOSTNAME": "test-postgresql",
+                "POSTGRESQL_DB_NAME": "django-k8s",
+                "POSTGRESQL_DB_NETLOC": "test-username:test-password@test-postgresql:5432",
+                "POSTGRESQL_DB_PARAMS": "",
+                "POSTGRESQL_DB_PASSWORD": "test-password",
+                "POSTGRESQL_DB_PATH": "/django-k8s",
+                "POSTGRESQL_DB_PORT": "5432",
+                "POSTGRESQL_DB_QUERY": "",
+                "POSTGRESQL_DB_SCHEME": "postgresql",
+                "POSTGRESQL_DB_USERNAME": "test-username",
+            },
+            id="django-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            "spring-boot",
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            'bash -c "java -jar *.jar"',
+            {
+                "APP_BASE_URL": "http://juju.test/",
+                "APP_METRICS_PORT": "8080",
+                "APP_OIDC_REDIRECT_PATH": "/oauth/callback",
+                "APP_OIDC_SCOPES": "openid profile email phone",
+                "APP_OIDC_USER_NAME_ATTRIBUTE": "email",
+                "APP_PEER_FQDNS": "spring-boot-k8s-0.spring-boot-k8s-endpoints.test-model.svc.cluster.local",
+                "APP_SECRET_KEY": "test",
+                "MANAGEMENT_SERVER_PORT": "8080",
+                "METRICS_PATH": "/actuator/prometheus",
+                "OTEL_LOGS_EXPORTER": "none",
+                "OTEL_METRICS_EXPORTER": "none",
+                "OTEL_TRACES_EXPORTER": "none",
+                "POSTGRESQL_DB_CONNECT_STRING": "postgresql://test-username:test-password@test-postgresql:5432/spring-boot-k8s",
+                "POSTGRESQL_DB_FRAGMENT": "",
+                "POSTGRESQL_DB_HOSTNAME": "test-postgresql",
+                "POSTGRESQL_DB_NAME": "spring-boot-k8s",
+                "POSTGRESQL_DB_NETLOC": "test-username:test-password@test-postgresql:5432",
+                "POSTGRESQL_DB_PARAMS": "",
+                "POSTGRESQL_DB_PASSWORD": "test-password",
+                "POSTGRESQL_DB_PATH": "/spring-boot-k8s",
+                "POSTGRESQL_DB_PORT": "5432",
+                "POSTGRESQL_DB_QUERY": "",
+                "POSTGRESQL_DB_SCHEME": "postgresql",
+                "POSTGRESQL_DB_USERNAME": "test-username",
+                "SERVER_PORT": "8080",
+                "management.endpoints.web.base-path": "/actuator",
+                "management.endpoints.web.exposure.include": "prometheus",
+                "management.endpoints.web.path-mapping.prometheus": "prometheus",
+                "server.forward-headers-strategy": "framework",
+                "spring.datasource.password": "test-password",
+                "spring.datasource.url": "jdbc:postgresql://test-postgresql:5432/spring-boot-k8s",
+                "spring.datasource.username": "test-username",
+                "spring.jpa.hibernate.ddl-auto": "none",
+                "spring.security.oauth2.client.provider.oidc.authorization-uri": "https://traefik_ip/model_name-hydra/oauth2/auth",
+                "spring.security.oauth2.client.provider.oidc.issuer-uri": "https://traefik_ip/model_name-hydra",
+                "spring.security.oauth2.client.provider.oidc.jwk-set-uri": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
+                "spring.security.oauth2.client.provider.oidc.token-uri": "https://traefik_ip/model_name-hydra/oauth2/token",
+                "spring.security.oauth2.client.provider.oidc.user-info-uri": "https://traefik_ip/model_name-hydra/userinfo",
+                "spring.security.oauth2.client.provider.oidc.user-name-attribute": "email",
+                "spring.security.oauth2.client.registration.oidc.authorization-grant-type": "authorization_code",
+                "spring.security.oauth2.client.registration.oidc.client-id": "test-client-id",
+                "spring.security.oauth2.client.registration.oidc.client-secret": "abc",
+                "spring.security.oauth2.client.registration.oidc.redirect-uri": "http://juju.test//oauth/callback",
+                "spring.security.oauth2.client.registration.oidc.scope": "openid,profile,email,phone",
+                "spring.security.oauth2.client.registration.oidc.user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
         ),
     ],
 )
@@ -105,8 +205,8 @@ def test_oauth_config_wrong_relation_order(
             FlaskCharm,
             "flask",
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "openid profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
             },
             "/bin/python3 -m gunicorn -c /flask/gunicorn.conf.py app:app -k [ sync ]",
             {
@@ -126,6 +226,104 @@ def test_oauth_config_wrong_relation_order(
                 "FLASK_OIDC_JWKS_URL": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
             },
             id="flask-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            "spring-boot",
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            'bash -c "java -jar *.jar"',
+            {
+                "APP_BASE_URL": "http://juju.test/",
+                "APP_METRICS_PORT": "8080",
+                "APP_OIDC_REDIRECT_PATH": "/oauth/callback",
+                "APP_OIDC_SCOPES": "openid profile email phone",
+                "APP_OIDC_USER_NAME_ATTRIBUTE": "email",
+                "APP_PEER_FQDNS": "spring-boot-k8s-0.spring-boot-k8s-endpoints.test-model.svc.cluster.local",
+                "APP_SECRET_KEY": "test",
+                "MANAGEMENT_SERVER_PORT": "8080",
+                "METRICS_PATH": "/actuator/prometheus",
+                "OTEL_LOGS_EXPORTER": "none",
+                "OTEL_METRICS_EXPORTER": "none",
+                "OTEL_TRACES_EXPORTER": "none",
+                "POSTGRESQL_DB_CONNECT_STRING": "postgresql://test-username:test-password@test-postgresql:5432/spring-boot-k8s",
+                "POSTGRESQL_DB_FRAGMENT": "",
+                "POSTGRESQL_DB_HOSTNAME": "test-postgresql",
+                "POSTGRESQL_DB_NAME": "spring-boot-k8s",
+                "POSTGRESQL_DB_NETLOC": "test-username:test-password@test-postgresql:5432",
+                "POSTGRESQL_DB_PARAMS": "",
+                "POSTGRESQL_DB_PASSWORD": "test-password",
+                "POSTGRESQL_DB_PATH": "/spring-boot-k8s",
+                "POSTGRESQL_DB_PORT": "5432",
+                "POSTGRESQL_DB_QUERY": "",
+                "POSTGRESQL_DB_SCHEME": "postgresql",
+                "POSTGRESQL_DB_USERNAME": "test-username",
+                "SERVER_PORT": "8080",
+                "management.endpoints.web.base-path": "/actuator",
+                "management.endpoints.web.exposure.include": "prometheus",
+                "management.endpoints.web.path-mapping.prometheus": "prometheus",
+                "server.forward-headers-strategy": "framework",
+                "spring.datasource.password": "test-password",
+                "spring.datasource.url": "jdbc:postgresql://test-postgresql:5432/spring-boot-k8s",
+                "spring.datasource.username": "test-username",
+                "spring.jpa.hibernate.ddl-auto": "none",
+                "spring.security.oauth2.client.provider.oidc.authorization-uri": "https://traefik_ip/model_name-hydra/oauth2/auth",
+                "spring.security.oauth2.client.provider.oidc.issuer-uri": "https://traefik_ip/model_name-hydra",
+                "spring.security.oauth2.client.provider.oidc.jwk-set-uri": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
+                "spring.security.oauth2.client.provider.oidc.token-uri": "https://traefik_ip/model_name-hydra/oauth2/token",
+                "spring.security.oauth2.client.provider.oidc.user-info-uri": "https://traefik_ip/model_name-hydra/userinfo",
+                "spring.security.oauth2.client.provider.oidc.user-name-attribute": "email",
+                "spring.security.oauth2.client.registration.oidc.authorization-grant-type": "authorization_code",
+                "spring.security.oauth2.client.registration.oidc.client-id": "test-client-id",
+                "spring.security.oauth2.client.registration.oidc.client-secret": "abc",
+                "spring.security.oauth2.client.registration.oidc.redirect-uri": "http://juju.test//oauth/callback",
+                "spring.security.oauth2.client.registration.oidc.scope": "openid,profile,email,phone",
+                "spring.security.oauth2.client.registration.oidc.user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            "django",
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            "/bin/python3 -m gunicorn -c /django/gunicorn.conf.py django_app.wsgi:application -k [ sync ]",
+            {
+                "DJANGO_OIDC_REDIRECT_PATH": "/oauth/callback",
+                "DJANGO_OIDC_SCOPES": "openid profile email phone",
+                "DJANGO_BASE_URL": "http://juju.test/",
+                "DJANGO_SECRET_KEY": "test",
+                "DJANGO_ALLOWED_HOSTS": '["juju.test"]',
+                "DJANGO_PEER_FQDNS": "django-k8s-0.django-k8s-endpoints.test-model.svc.cluster.local",
+                "DJANGO_OIDC_CLIENT_ID": "test-client-id",
+                "DJANGO_OIDC_CLIENT_SECRET": "abc",
+                "DJANGO_OIDC_API_BASE_URL": "https://traefik_ip/model_name-hydra",
+                "DJANGO_OIDC_AUTHORIZE_URL": "https://traefik_ip/model_name-hydra/oauth2/auth",
+                "DJANGO_OIDC_ACCESS_TOKEN_URL": "https://traefik_ip/model_name-hydra/oauth2/token",
+                "DJANGO_OIDC_USER_URL": "https://traefik_ip/model_name-hydra/userinfo",
+                "DJANGO_OIDC_CLIENT_KWARGS": '{"scope": "openid profile email phone"}',
+                "DJANGO_OIDC_JWKS_URL": "https://traefik_ip/model_name-hydra/.well-known/jwks.json",
+                "POSTGRESQL_DB_CONNECT_STRING": "postgresql://test-username:test-password@test-postgresql:5432/django-k8s",
+                "POSTGRESQL_DB_FRAGMENT": "",
+                "POSTGRESQL_DB_HOSTNAME": "test-postgresql",
+                "POSTGRESQL_DB_NAME": "django-k8s",
+                "POSTGRESQL_DB_NETLOC": "test-username:test-password@test-postgresql:5432",
+                "POSTGRESQL_DB_PARAMS": "",
+                "POSTGRESQL_DB_PASSWORD": "test-password",
+                "POSTGRESQL_DB_PATH": "/django-k8s",
+                "POSTGRESQL_DB_PORT": "5432",
+                "POSTGRESQL_DB_QUERY": "",
+                "POSTGRESQL_DB_SCHEME": "postgresql",
+                "POSTGRESQL_DB_USERNAME": "test-username",
+            },
+            id="django-oidc",
         ),
     ],
 )
@@ -179,10 +377,29 @@ def test_oauth_config_correct_relation_order(
             "flask_base_state",
             FlaskCharm,
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "openid profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
             },
             id="flask-oidc",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            id="django-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
         ),
     ],
 )
@@ -236,10 +453,29 @@ def test_oauth_config_remove_ingress_integration_should_block(
             "flask_base_state",
             FlaskCharm,
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "openid profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
             },
             id="flask-oidc",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            id="django-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
         ),
     ],
 )
@@ -291,10 +527,29 @@ def test_oauth_config_remove_oauth_integration_should_not_block(
             "flask_base_state",
             FlaskCharm,
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "profile email phone",
             },
             id="flask-oidc-fail",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "profile email phone",
+            },
+            id="django-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
         ),
     ],
 )
@@ -331,19 +586,106 @@ def test_oauth_config_wrong_scope(base_state: dict, charm, config: dict, request
 
 
 @pytest.mark.parametrize(
+    "base_state, charm, config",
+    [
+        pytest.param(
+            "flask_base_state",
+            FlaskCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            id="flask-oidc",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+            },
+            id="django-oidc",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+            },
+            id="spring-boot-oidc",
+        ),
+    ],
+)
+def test_blocked_when_relation_data_empty(base_state: dict, charm, config: dict, request) -> None:
+    """
+    arrange: set the workload charm config.
+    act: start the workload charm and integrate with oauth and ingress using wrong scope.
+    assert: workload charm should be blocked.
+    """
+    base_state = request.getfixturevalue(base_state)
+    base_state["config"] = config
+    secret_id = token_hex(16)
+    oauth_relation = testing.Relation(
+        endpoint="oidc", interface="oauth", remote_app_data={}, remote_app_name="OIDC_CHARM"
+    )
+    base_state["relations"].append(oauth_relation)
+    base_state["secrets"] = [testing.Secret(id=secret_id, tracked_content={"secret": "abc"})]
+    ingress_relation = testing.Relation(
+        endpoint="ingress",
+        interface="ingress",
+        remote_app_data={"ingress": '{"url": "http://juju.test/"}'},
+    )
+    base_state["relations"].append(ingress_relation)
+    state = testing.State(**base_state)
+    context = testing.Context(
+        charm_type=charm,
+    )
+    out = context.run(context.on.relation_changed(ingress_relation), state)
+    assert out.unit_status == testing.BlockedStatus("Please check OIDC_CHARM charm!")
+
+
+@pytest.mark.parametrize(
     "base_state, charm, config, multiple_oauth_integrations",
     [
         pytest.param(
             "flask_base_state",
             FlaskCharm,
             {
-                "oidc_redirect_path": "/oauth/callback",
-                "oidc_scopes": "openid profile email phone",
-                "google_redirect_path": "/oauth/callback",
-                "google_scopes": "openid profile email phone",
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "google-redirect-path": "/oauth/callback",
+                "google-scopes": "openid profile email phone",
             },
             {"framework": "flask"},
-            id="flask-oidc-fail",
+            id="flask",
+        ),
+        pytest.param(
+            "django_base_state",
+            DjangoCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "google-redirect-path": "/oauth/callback",
+                "google-scopes": "openid profile email phone",
+            },
+            {"framework": "django"},
+            id="django",
+        ),
+        pytest.param(
+            "spring_boot_base_state",
+            SpringBootCharm,
+            {
+                "oidc-redirect-path": "/oauth/callback",
+                "oidc-scopes": "openid profile email phone",
+                "oidc-user-name-attribute": "email",
+                "google-redirect-path": "/oauth/callback",
+                "google-scopes": "openid profile email phone",
+                "google-user-name-attribute": "email",
+            },
+            {"framework": "springboot"},
+            id="spring-boot",
         ),
     ],
     indirect=["multiple_oauth_integrations"],